# Speed up scans of Uncompressed strings