--- a/src/include/duckdb/storage/checkpoint/string_checkpoint_state.hpp
+++ b/src/include/duckdb/storage/checkpoint/string_checkpoint_state.hpp
@@ -33,19 +33,6 @@ struct StringBlock {
 	unique_ptr<StringBlock> next;
 };
 
-struct string_location_t { // NOLINT
-	string_location_t(block_id_t block_id, int32_t offset) : block_id(block_id), offset(offset) {
-	}
-	string_location_t() {
-	}
-	bool IsValid(const idx_t block_size) {
-		auto cast_block_size = NumericCast<int32_t>(block_size);
-		return offset < cast_block_size && (block_id == INVALID_BLOCK || block_id >= MAXIMUM_BLOCK);
-	}
-	block_id_t block_id;
-	int32_t offset;
-};
-
 struct UncompressedStringSegmentState : public CompressedSegmentState {
 	~UncompressedStringSegmentState() override;