This PR reworks the scan code of uncompressed strings, simplifying it and speeding it up.

Running the following query on a TPC-H SF10 database (in uncompressed format) leads to the following performance improvement:

```sql
select avg(strlen(l_comment)) from lineitem;
```

|  main  |  New  |
|--------|-------|
| 0.042s | 0.03s |