@carlopi 2024-11-29
One fun thing could be at some point being able to push down the fact that we want only the length of the string (at least for this query), so that `FetchStringFromDict` can be skipped entirely.